feat: 코드 분리
